Refactor GradientCollapse to new component
