+++ react/GradientCollapse.js
@@ -1,0 +1,3 @@
+import GradientCollapse from './components/GradientCollapse/index'
+
+export default GradientCollapse
