+++ react/components/GradientCollapse/styles.css
@@ -1,0 +1,15 @@
+/**
+*   GradientCollapse CSS Modules
+*/
+
+.fadeBottom {
+    background-image: linear-gradient(transparent, white);
+}
+
+.pointerEventsNone {
+    pointer-events: none;
+}
+
+.pointerEventsAuto {
+    pointer-events: auto;
+}
